Phase II Core Programming Updates/Fixes
I have updated the boss and cleaned up the code as well as the enemies and the ending of the game has some new things added on such as buttons to navigate and fly away player. Lastley but not least is the new quit button at the main menu to exit the game!